make <user> a component
