+++ app/components/user/template.html
@@ -1,4 +1,4 @@
-<div ng-controller="UserCtrl">
+<div>
   <div class="ph-header ph-header-large ph-lg-margin-bottom">
     <div class="container">
       <div class="row">
@@ -52,7 +52,6 @@
         </div>
       </div>
     </div>
-
   </div>
   <div class="container">
     <div app-view-segment="1"></div>
